chore: Move asset test to own mod
